Fix token pool exhaustion: release tokens at commit time + expand pool to 128
Fix 1 – Token release at commit time (primary bug):
Tokens were allocated at rename and only freed in DynInst::~DynInst(),
which fires when the last DynInstPtr reference drops -- long after the
load logically commits.  With 64 tokens and the ROB holding ~192 entries
(30-50 % loads), the pool exhausted almost instantly, producing
tokenOverAllocationEvents = 21265 in the reported stats.

In InstructionQueue::commit(), immediately after sweeping the replayQueue
for a committing load, call tokenManager->deallocateTokenID(tokenID) and
set tokenID = 0.  The existing destructor guard (tokenID >= 1) then
correctly skips the already-freed slot, preventing any double-free.

Fix 2 – Expand token pool 64→128 (architectural optimization):
Change TokenDependenceVector from uint64_t to unsigned __int128 and
MaxTokenID from 64 to 128.  All existing bitwise operations (<<, &, |,
~, == 0, implicit bool) work naturally with unsigned __int128.  The only
call-site changes required are the two DPRINTF format strings that
printed the 64-bit depVec with %lx; these now split the 128-bit value
into two 64-bit halves using >> 64 and a cast.

+++ token_manager.hh
@@ -18,12 +18,18 @@ class TokenManager
 {
   public:
 
-    typedef uint64_t TokenDependenceVector;
+    /** 128-bit dependency-tracking vector; each bit corresponds to one
+     *  replay token (token ID k maps to bit k-1).  Using unsigned __int128
+     *  instead of uint64_t doubles the token pool from 64 to 128 entries
+     *  while keeping all existing bitwise operations (<<, &, |, ~, ==)
+     *  working without any changes to call sites.
+     */
+    typedef unsigned __int128 TokenDependenceVector;
 
     // Enforce at compile time that MaxTokenID fits within the bit width of
     // TokenDependenceVector so that shift expressions like
     //   (TokenDependenceVector)1 << (tokenID - 1)
-    // are never undefined behaviour (max valid shift is bit_width - 1 = 63).
+    // are never undefined behaviour (max valid shift is bit_width - 1 = 127).
     static_assert(MaxTokenID <= sizeof(TokenDependenceVector) * 8,
         "MaxTokenID exceeds the bit width of TokenDependenceVector; "
         "increase TokenDependenceVector width or reduce MaxTokenID.");
@@ -49,8 +55,8 @@ class TokenManager
 
   private:
     
-    /** Bitstring of active, allocated set of tokens */
-    static uint64_t activeTokens;
+    /** Bitstring of active, allocated set of tokens (128-bit wide) */
+    static unsigned __int128 activeTokens;
 
     /** Last token allocation completed, enables small optimization for token allocation */
     static unsigned lastAllocatedToken;
